backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
torch>=2.1.0
torchvision>=0.16.0
timm==1.0.3
opencv-python-headless==4.9.0.80
ultralytics==8.2.0
Pillow>=10.0.0
numpy>=1.24.0
